Distributed-Systems-main/consensus/requirements.txt
grpcio==1.76.0
protobuf==6.31.1


